♻️ refactor(agent): hoist __mainEditor exposure to the conversation level
MainChatInput is only rendered for homogeneous agents — ConversationArea swaps in
HeterogeneousChatInput for Claude Code / Codex style agents, so window.__mainEditor
stayed undefined on exactly the pages automation most wants to drive.

Move the mount into ExposeMainEditor, rendered inside ConversationProvider so both
composers are covered. It stays out of the shared features/Conversation/ChatInput on
purpose: AgentBuilder and FloatingChatPanel render that same input and would otherwise
claim the handle.
